test(gui): fold entity CRUD suites into manifest-listed test files
The three entity CRUD vitest files were never registered in the protected
GUI test manifest, so npm run test:gui failed manifest validation on the
final commit. Merge their coverage into the adjacent manifest-listed
suites - pure preview derivation checks into entity-locator-renderer,
wizard / directory browser / detail action / kind form checks into
entities-view - and drop the unregistered files.

Also replace single-tick waits with condition polling where assertions
depend on an async query chain (sidebar connection endpoint, wizard
preview, post-import body, directory file body): full-suite parallel
load regularly outlasts one microtask tick and turned those checks
flaky.
