+++ packages/gui/test/entities-view.vitest.ts
@@ -5,8 +5,16 @@ import { act } from "react";
 import { createRoot, type Root } from "react-dom/client";
 import { createElement } from "react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { createHash } from "node:crypto";
 import { EntitiesView } from "../src/renderer/views/EntitiesView.tsx";
 import { CURATED_ENTITY_DOC_GROUPS, FACT_TYPE_VOCABULARY } from "../src/renderer/entity-docs.ts";
+import { sourceIdentityOf } from "../src/renderer/entity-import-preview.ts";
+import {
+  VerticalKindForm,
+  describeRelationsIssue,
+  locateJsonError,
+} from "../src/renderer/components/entityDoc/VerticalKindForm.tsx";
+import type { ArtifactKindDeclaration } from "../src/renderer/vertical-kind-client.ts";
 import { setActiveLocale } from "../src/renderer/i18n/core.ts";
 import type { ViewId } from "../src/renderer/navigation/viewHistory.ts";
 
@@ -69,14 +77,23 @@ function declaredAdrKindRow(overrides: Record<string, unknown> = {}) {
   };
 }
 
-function governedRow(entityId: string, title: string | null = null) {
+function governedRow(
+  entityId: string,
+  title: string | null = null,
+  overrides: {
+    readonly locator?: string;
+    readonly revision?: number;
+    readonly archived?: boolean;
+  } = {},
+) {
   return {
     kind: ADR_KIND,
     entityId,
     ref: `${ADR_KIND}/${entityId}`,
     title,
-    locator: { kind: "repository-path", value: `docs/adr/${entityId}.md` },
-    revision: 0,
+    locator: { kind: "repository-path", value: overrides.locator ?? `docs/adr/${entityId}.md` },
+    revision: overrides.revision ?? 0,
+    ...(overrides.archived === undefined ? {} : { archived: overrides.archived }),
   };
 }
 
@@ -566,5 +583,479 @@ describe("catalog-page entity list (goal 4)", () => {
     const emptyCatalog = await renderSurface(view(null));
     await settle();
     expect(emptyCatalog.querySelector('[data-testid="governed-entity-catalog-list"]')).toBeNull();
+  });
+});
+
+/**
+ * 实体 CRUD 判据(task_a494eac2 Goal 1/2/4):新建向导从既有实体的公共父目录起步,
+ * 浏览/选定/预览(推导 id 与 kernel 公式一致)后走 repo.entity.import
+ * (expectedVersion 恒 0,title 留空时不发);目录 locator 的树懒展开(展开才对子
+ * 路径发读),点文件在内嵌查看器渲染;编辑/归档在详情操作区,列表行不挂写动作。
+ * locator 读面按文件内容/目录条目两张固定表回包,写路径回执全部落在 state 里。
+ */
+
+const FILE_CONTENT: Readonly<Record<string, string>> = {
+  "docs/adr/ADR-0001.md": "# ADR-0001 · 探针\n\n这条正文来自 locator 读面。",
+  "docs/adr/ADR-0002.md": "# ADR-0002 · 复核\n\n第二条正文。",
+  "docs/adr/ADR-0003.md": "# 新导入的研究\n\n导入后的正文。",
+  "docs/adr/research/README.md": "# 研究包说明\n\n目录正文。",
+  "docs/adr/research/notes.md": "# 笔记\n\n目录里的文件。",
+};
+
+const DIRECTORY_ENTRIES: Readonly<Record<string, ReadonlyArray<{ path: string; directory: boolean }>>> = {
+  "docs/adr": [
+    { path: "docs/adr/ADR-0001.md", directory: false },
+    { path: "docs/adr/ADR-0002.md", directory: false },
+    { path: "docs/adr/research", directory: true },
+  ],
+  "docs/adr/research": [
+    { path: "docs/adr/research/README.md", directory: false },
+    { path: "docs/adr/research/notes.md", directory: false },
+  ],
+};
+
+function derivedId(path: string): string {
+  return `ADR-${createHash("sha256").update(sourceIdentityOf(REPO_ID, path)).digest("hex").slice(0, 16)}`;
+}
+
+interface CrudBridgeState {
+  rows: ReturnType<typeof governedRow>[];
+  locatorCalls: string[];
+  imports: unknown[];
+  updates: unknown[];
+  archives: unknown[];
+}
+
+function crudRow(entityId: string, locator: string): ReturnType<typeof governedRow> {
+  return governedRow(entityId, null, { locator, revision: 4 });
+}
+
+function stubCrudBridge(initialRows: ReturnType<typeof governedRow>[] = []): CrudBridgeState {
+  const state: CrudBridgeState = { rows: [...initialRows], locatorCalls: [], imports: [], updates: [], archives: [] };
+  vi.stubGlobal("window", {
+    harness: {
+      getWorkspaceSummary: vi.fn(async () => ({
+        schema: "daemon.workspace-summary/v1",
+        ok: true,
+        status: "ready",
+        tasks: { total: 0, byStatus: {} },
+        decisions: { total: 0, inboxCount: 0, byState: {}, groups: [] },
+        watermark: 0,
+        sourceRevision: 0,
+      })),
+      getCatalogSnapshot: vi.fn(async () => ({
+        schema: "gui-catalog-snapshot/v1",
+        ok: true,
+        status: "ready",
+        repoId: REPO_ID,
+        observedAt: "2026-09-06T00:00:00.000Z",
+        catalogDigest: "digest000000000000",
+        defaults: {},
+        presets: [],
+        verticals: [],
+        templates: [],
+        scaffolds: { task: [], repository: [] },
+        adapters: [],
+      })),
+      listAgents: vi.fn(async () => ({ schema: "agent-entity-catalog/v1", ok: true, agents: [] })),
+      listSquads: vi.fn(async () => ({ schema: "squad-entity-catalog/v1", ok: true, squads: [] })),
+      listSchedules: vi.fn(async () => ({ ok: true, status: "ready", schedules: [], watermark: 0, sourceRevision: 0 })),
+      getRelationGraph: vi.fn(async () => ({
+        ok: true,
+        facet: "facts",
+        edges: [],
+        coverageRows: [],
+        factAnchors: [],
+        facts: [],
+        domainTypes: [],
+        warnings: [],
+      })),
+      readEntityKinds: vi.fn(async () => ({ schema: "entity-kind-catalog/v1", kinds: [declaredAdrKindRow()] })),
+      readEntityRows: vi.fn(async () => ({ schema: "entity-row-list/v1", ok: true, rows: state.rows })),
+      readVerticalDeclaration: vi.fn(async () => ({
+        schema: "repository-vertical-declaration-read/v1",
+        declarationRevision: 7,
+        declaration: { entityKinds: [] },
+      })),
+      readEntityLocator: vi.fn(async ({ locatorValue }: { readonly locatorValue: string }) => {
+        state.locatorCalls.push(locatorValue);
+        const files = FILE_ENTRIES_PASSTHROUGH(locatorValue);
+        if (files !== null) return files;
+        const entries = DIRECTORY_ENTRIES[locatorValue];
+        if (entries)
+          return {
+            schema: "entity-locator-read/v1",
+            outcome: "directory",
+            path: locatorValue,
+            content: null,
+            sizeBytes: null,
+            entries,
+            truncated: false,
+          };
+        return {
+          schema: "entity-locator-read/v1",
+          outcome: "missing",
+          path: locatorValue,
+          content: null,
+          sizeBytes: null,
+          entries: [],
+          truncated: false,
+        };
+      }),
+      importEntity: vi.fn(async (payload: object) => {
+        state.imports.push(payload);
+        const locator = (payload as { locator: string }).locator;
+        state.rows.push(crudRow(derivedId(locator), locator));
+        return {
+          schema: "command-receipt/v2",
+          ok: true,
+          command: "entity.import",
+          outcome: "applied",
+          opId: "op-import-1",
+        };
+      }),
+      updateEntity: vi.fn(async (payload: object) => {
+        state.updates.push(payload);
+        return {
+          schema: "command-receipt/v2",
+          ok: true,
+          command: "entity.update",
+          outcome: "applied",
+          opId: "op-update-1",
+        };
+      }),
+      archiveEntity: vi.fn(async (payload: object) => {
+        state.archives.push(payload);
+        return {
+          schema: "command-receipt/v2",
+          ok: true,
+          command: "entity.archive",
+          outcome: "applied",
+          opId: "op-archive-1",
+        };
+      }),
+    },
+  });
+  return state;
+}
+
+function FILE_ENTRIES_PASSTHROUGH(locatorValue: string) {
+  const content = FILE_CONTENT[locatorValue];
+  if (content === undefined) return null;
+  return {
+    schema: "entity-locator-read/v1",
+    outcome: "file",
+    path: locatorValue,
+    content,
+    sizeBytes: content.length,
+    entries: [],
+    truncated: false,
+  };
+}
+
+async function renderCrudView(focusedRef: string | null): Promise<HTMLElement> {
+  const container = await renderSurface(view(focusedRef));
+  await settle();
+  return container;
+}
+
+async function click(container: HTMLElement, testId: string): Promise<void> {
+  const button = container.querySelector<HTMLButtonElement>(`[data-testid="${testId}"]`);
+  expect(button, testId).not.toBeNull();
+  await act(async () => {
+    button!.click();
+  });
+  await settle();
+}
+
+describe("new entity wizard (goal 1)", () => {
+  it("starts the browser at the common parent of existing locators", async () => {
+    stubCrudBridge([crudRow("ADR-0001", "docs/adr/ADR-0001.md"), crudRow("ADR-0002", "docs/adr/ADR-0002.md")]);
+    const container = await renderCrudView(`entitydoc/${ADR_KIND}`);
+    await click(container, "governed-entity-new");
+    expect(container.querySelector('[data-testid="new-entity-wizard"]')).not.toBeNull();
+    // seed = 既有 locator 的最深公共父目录,不是手输,也不是仓根。
+    expect(container.querySelector('[data-testid="repo-path-browser-location"]')?.textContent).toBe("docs/adr");
+    expect(container.querySelector('[data-testid="new-entity-wizard-seed"]')).toBeNull();
+  });
+
+  it("asks for a seed only when no existing entity can suggest one", async () => {
+    stubCrudBridge();
+    const container = await renderCrudView(`entitydoc/${ADR_KIND}`);
+    await click(container, "governed-entity-new");
+    expect(container.querySelector('[data-testid="new-entity-wizard-seed"]')).not.toBeNull();
+    expect(container.querySelector('[data-testid="repo-path-browser"]')).toBeNull();
+  });
+
+  it("navigates directories, previews the derived title and id, then imports without a hand-written id", async () => {
+    const state = stubCrudBridge([crudRow("ADR-0001", "docs/adr/ADR-0001.md")]);
+    const container = await renderCrudView(`entitydoc/${ADR_KIND}`);
+    await click(container, "governed-entity-new");
+    // 进入子目录:对子路径再发同一条目录读。
+    await click(container, "repo-path-entry-docs/adr/research");
+    expect(state.locatorCalls).toContain("docs/adr/research");
+    expect(container.querySelector('[data-testid="repo-path-browser-location"]')?.textContent).toBe(
+      "docs/adr/research",
+    );
+    // 选定文件:预览给出推导 title(文件首标题)与推导 id(kernel 公式)。
+    await click(container, "repo-path-entry-docs/adr/research/notes.md");
+    // 预览要等两条 query(内容读 + WebCrypto id 推导)都落定,轮询到内容就位。
+    await vi.waitFor(() =>
+      expect(container.querySelector('[data-testid="new-entity-wizard-preview-area"]')?.textContent).toContain("笔记"),
+    );
+    const preview = container.querySelector('[data-testid="new-entity-wizard-preview-area"]');
+    expect(preview?.textContent).toContain(derivedId("docs/adr/research/notes.md"));
+    // 导入:expectedVersion 恒 0,title 留空就不发,人不填任何身份字段。
+    await click(container, "new-entity-wizard-submit");
+    expect(state.imports).toEqual([
+      { repoId: REPO_ID, entityKind: ADR_KIND, locator: "docs/adr/research/notes.md", expectedVersion: 0 },
+    ]);
+    // 回执 applied 后:向导退出,行缓存刷新,新实体被选中并渲染正文。
+    expect(container.querySelector('[data-testid="new-entity-wizard"]')).toBeNull();
+    await vi.waitFor(() =>
+      expect(container.querySelector('[data-testid="entity-locator-markdown"]')?.textContent).toContain("目录里的文件"),
+    );
+  });
+
+  it("imports a directory target with the README heading as the derived title", async () => {
+    const state = stubCrudBridge([crudRow("ADR-0001", "docs/adr/ADR-0001.md")]);
+    const container = await renderCrudView(`entitydoc/${ADR_KIND}`);
+    await click(container, "governed-entity-new");
+    await click(container, "repo-path-entry-docs/adr/research");
+    await click(container, "repo-path-browser-pick-directory");
+    // 目录目标要等 listing → README 读的串行链加 id 推导,轮询到内容就位。
+    await vi.waitFor(() =>
+      expect(container.querySelector('[data-testid="new-entity-wizard-preview-area"]')?.textContent).toContain(
+        "研究包说明",
+      ),
+    );
+    await click(container, "new-entity-wizard-submit");
+    expect(state.imports).toEqual([
+      { repoId: REPO_ID, entityKind: ADR_KIND, locator: "docs/adr/research", expectedVersion: 0 },
+    ]);
+  });
+});
+
+describe("directory locator browser (goal 2)", () => {
+  it("lazily expands subdirectories and opens files in the inline viewer", async () => {
+    const state = stubCrudBridge([crudRow("ADR-0001", "docs/adr")]);
+    const container = await renderCrudView(`${ADR_KIND}/ADR-0001`);
+    // 树根钉在实体目录:直接看到它的条目,不再是全路径分段的两级壳。
+    const tree = container.querySelector('[data-testid="entity-locator-directory"]');
+    expect(tree).not.toBeNull();
+    expect(tree?.textContent).toContain("research/");
+    expect(tree?.textContent).toContain("ADR-0002.md");
+    // 未展开前不对子目录发读。
+    expect(state.locatorCalls).not.toContain("docs/adr/research");
+    await click(container, "entity-directory-node-docs/adr/research");
+    expect(state.locatorCalls).toContain("docs/adr/research");
+    // 点文件:内嵌查看器按渲染器选择表渲染 markdown(内容读是异步链,轮询到就位)。
+    await click(container, "entity-directory-node-docs/adr/research/notes.md");
+    await vi.waitFor(() =>
+      expect(container.querySelector('[data-testid="entity-directory-file-markdown"]')?.textContent).toContain(
+        "目录里的文件",
+      ),
+    );
+  });
+
+  it("shows the dedicated pdf card instead of pretending to render bytes", async () => {
+    stubCrudBridge([crudRow("ADR-0001", "docs/adr/ADR-0001.md")]);
+    const container = await renderCrudView(`${ADR_KIND}/ADR-0001`);
+    expect(container.querySelector('[data-testid="entity-locator-pdf"]')).toBeNull();
+    expect(container.querySelector('[data-testid="entity-locator-markdown"]')).not.toBeNull();
+  });
+});
+
+describe("detail action area (goal 4)", () => {
+  it("keeps edit and archive out of the list rows and acts on the selected entity", async () => {
+    const state = stubCrudBridge([crudRow("ADR-0001", "docs/adr/ADR-0001.md")]);
+    const container = await renderCrudView(`entitydoc/${ADR_KIND}`);
+    await click(container, "governed-entity-row-ADR-0001");
+    // 编辑/归档在详情操作区;列表行内不再出现这两个词。
+    const row = container.querySelector('[data-testid="governed-entity-row-ADR-0001"]');
+    expect(row?.textContent).not.toContain("编辑");
+    expect(container.querySelector('[data-testid="entity-detail-actions"]')).not.toBeNull();
+    // 编辑:保存走 repo.entity.update,expectedVersion 用行上的 revision。
+    await click(container, "entity-detail-edit");
+    const title = container.querySelector<HTMLInputElement>('input[aria-label="title"]');
+    await act(async () => {
+      const setter = Object.getOwnPropertyDescriptor(HTMLInputElement.prototype, "value")?.set;
+      setter?.call(title!, "新标题");
+      title!.dispatchEvent(new Event("input", { bubbles: true }));
+    });
+    await click(container, "entity-detail-edit-save");
+    expect(state.updates[0]).toMatchObject({
+      repoId: REPO_ID,
+      entityKind: ADR_KIND,
+      entityId: "ADR-0001",
+      expectedVersion: 4,
+      title: "新标题",
+      locator: "docs/adr/ADR-0001.md",
+    });
+    // 归档:带原因走 repo.entity.archive。
+    await click(container, "entity-detail-archive");
+    const reason = container.querySelector<HTMLInputElement>('input[aria-label="归档原因"]');
+    await act(async () => {
+      const setter = Object.getOwnPropertyDescriptor(HTMLInputElement.prototype, "value")?.set;
+      setter?.call(reason!, "过期");
+      reason!.dispatchEvent(new Event("input", { bubbles: true }));
+    });
+    await click(container, "entity-detail-archive-confirm");
+    expect(state.archives[0]).toMatchObject({
+      repoId: REPO_ID,
+      entityId: "ADR-0001",
+      expectedVersion: 4,
+      reason: "过期",
+    });
+  });
+});
+
+/**
+ * kind 声明表单的收敛判据(task_a494eac2 Goal 3):编辑时身份/存储字段只读展示
+ * (没有任何输入框),可编辑的只有 display / maturityVocabulary / relations;新建时
+ * 按模板预填并逐字段说明;relations 的 JSON 错误能定位到行列或下标。
+ */
+
+const KIND_FORM_INITIAL: ArtifactKindDeclaration = {
+  id: "architecture-decision-record",
+  entityType: "artifact",
+  version: 2,
+  idPrefix: "ADR",
+  display: { singular: "Architecture Decision Record", plural: "Architecture Decision Records" },
+  descriptorSchemaRef: "schema://artifact-descriptor",
+  store: { pathTemplate: "entities/adrs/{id}.json" },
+  locatorKinds: ["repository-path"],
+  maturityVocabulary: ["draft", "accepted"],
+};
+
+async function renderKindForm(initial?: ArtifactKindDeclaration): Promise<HTMLElement> {
+  return renderSurface(
+    createElement(VerticalKindForm, {
+      ...(initial === undefined ? {} : { initial }),
+      busy: false,
+      error: null,
+      onCancel: noop,
+      onSubmit: noop,
+    }),
+  );
+}
+
+async function typeInto(container: HTMLElement, label: string, text: string): Promise<void> {
+  const input = container.querySelector<HTMLInputElement>(`input[aria-label="${label}"]`);
+  expect(input, label).not.toBeNull();
+  await act(async () => {
+    const setter = Object.getOwnPropertyDescriptor(HTMLInputElement.prototype, "value")?.set;
+    setter?.call(input!, text);
+    input!.dispatchEvent(new Event("input", { bubbles: true }));
+  });
+}
+
+async function typeTextarea(container: HTMLElement, label: string, text: string): Promise<void> {
+  const area = container.querySelector<HTMLTextAreaElement>(`textarea[aria-label="${label}"]`);
+  expect(area, label).not.toBeNull();
+  await act(async () => {
+    const setter = Object.getOwnPropertyDescriptor(HTMLTextAreaElement.prototype, "value")?.set;
+    setter?.call(area!, text);
+    area!.dispatchEvent(new Event("input", { bubbles: true }));
+  });
+}
+
+describe("vertical kind form edit mode: identity fields are read-only display", () => {
+  it("renders identity/storage as a definition list with no inputs for them", async () => {
+    const container = await renderKindForm(KIND_FORM_INITIAL);
+    const identity = container.querySelector('[data-testid="vertical-kind-identity"]');
+    expect(identity?.textContent).toContain("architecture-decision-record");
+    expect(identity?.textContent).toContain("entities/adrs/{id}.json");
+    for (const label of ["id", "version", "idPrefix", "descriptorSchemaRef", "store.pathTemplate"]) {
+      expect(container.querySelector(`input[aria-label="${label}"]`), label).toBeNull();
+      expect(container.querySelector(`input[aria-label="${label}"][disabled]`), label).toBeNull();
+    }
+    // 可编辑的仍在:display 与词表。
+    expect(container.querySelector('input[aria-label="display.singular"]')).not.toBeNull();
+    expect(container.querySelector('input[aria-label="maturityVocabulary(逗号分隔)"]')).not.toBeNull();
+  });
+});
+
+describe("vertical kind form new mode: template, not a blank form", () => {
+  it("prefills template defaults and keeps identity editable with hints", async () => {
+    const container = await renderKindForm();
+    expect(container.querySelector<HTMLInputElement>('input[aria-label="descriptorSchemaRef"]')?.value).toBe(
+      "schema://artifact-descriptor",
+    );
+    expect(container.querySelector<HTMLInputElement>('input[aria-label="store.pathTemplate"]')?.value).toBe(
+      "entities/{id}.json",
+    );
+    const locatorCheckbox = container.querySelector<HTMLInputElement>('input[type="checkbox"]');
+    expect(locatorCheckbox?.checked).toBe(true);
+    // 字段说明在场:不是空白表单让人盲填。
+    expect(container.textContent).toContain("创建后不可改");
+    expect(container.textContent).toContain("描述符落盘路径模板");
+  });
+});
+
+describe("vertical kind form relations JSON editor localizes errors", () => {
+  it("locates a parse error to line and column", () => {
+    const message = locateJsonError("Unexpected token '}', ...\"}\" is not valid JSON (line 2 column 5)", "[\n}]");
+    expect(message).toContain("第 2 行第 5 列");
+  });
+
+  it("converts a byte position into line and column", () => {
+    const message = locateJsonError("Unexpected end of JSON input at position 7", "[\n  {}");
+    expect(message).toContain("第 2 行");
+  });
+
+  it("reports the item index for structural violations", () => {
+    const relation = {
+      type: "relates",
+      sourceKind: "task",
+      targetKind: "software/coding/x@1",
+      reads: "读取理由",
+      strength: "weak",
+      decisionClaimRef: "decision/dec_X/C1",
+      decisionContentPin: `sha256:${"0".repeat(64)}`,
+    };
+    expect(describeRelationsIssue(JSON.stringify([relation]))).toBeNull();
+    expect(describeRelationsIssue("not json")).toContain("JSON 解析失败");
+    expect(describeRelationsIssue('{"a":1}')).toBe("relations 必须是 JSON 数组。");
+    expect(describeRelationsIssue(JSON.stringify([{ ...relation, strength: "medium" }]))).toContain(
+      "relations[0]:strength 只能是 weak 或 strong",
+    );
+    const { decisionClaimRef: _omitted, ...missingClaim } = relation;
+    expect(describeRelationsIssue(JSON.stringify([relation, missingClaim]))).toContain(
+      "relations[1]:缺少字段 decisionClaimRef",
+    );
+    expect(describeRelationsIssue(JSON.stringify([{ ...relation, extra: 1 }]))).toContain(
+      "relations[0]:未声明字段 extra",
+    );
+  });
+
+  it("shows the issue inline while typing", async () => {
+    const container = await renderKindForm();
+    // 展开可折叠编辑器。
+    const toggle = container.querySelector<HTMLButtonElement>('[data-testid="vertical-kind-relations"] > button');
+    await act(async () => {
+      toggle!.click();
+    });
+    await typeTextarea(container, "relations JSON", "[\n}");
+    expect(container.querySelector('[data-testid="vertical-kind-relations-issue"]')?.textContent).toContain(
+      "JSON 解析失败",
+    );
+  });
+});
+
+describe("vertical kind form submission", () => {
+  it("keeps the submit disabled until required editable fields are valid", async () => {
+    const container = await renderKindForm();
+    const submit = container.querySelector<HTMLButtonElement>('button[type="submit"]');
+    expect(submit?.disabled).toBe(true);
+    await typeInto(container, "display.singular", "Research Note");
+    await typeInto(container, "display.plural", "Research Notes");
+    await typeInto(container, "idPrefix", "RSRCH");
+    await typeInto(container, "id", "research-note");
+    expect(submit?.disabled).toBe(false);
+    await typeInto(container, "idPrefix", "1BAD");
+    expect(submit?.disabled).toBe(true);
+    await typeInto(container, "idPrefix", "RSRCH");
+    expect(submit?.disabled).toBe(false);
   });
 });
